Cyber3DMap-Streamlit/src/requirements.txt
streamlit==1.39.0
plotly==5.24.1
networkx==3.3
requests==2.32.3
torch==2.4.1
torch-geometric==2.5.3
pyyaml==6.0.2